requirements-python3810.txt
PyYAML==6.0.1
torch==1.13.1+computecanada
tqdm==4.66.2
xarray==2023.1.0   #2024.5.0
scipy==1.10.1   #1.13.1
numpy==1.24.2+computecanada
matplotlib==3.7.5   #3.9.1
torchdiffeq==0.2.4
geopandas==0.13.2   #1.0.1
cmcrameri==1.9
scikit-learn==1.3.2